Move mapper extensions to TvManiac kit.
