+++ ios/Modules/TvManiacUI/Sources/TvManiacUI/Models/SwiftShow.swift
@@ -1,5 +1,4 @@
 import Foundation
-import TvManiac
 
 public struct SwiftShow: Identifiable {
     public let id: UUID = .init()
@@ -25,17 +24,4 @@ public struct SwiftShow: Identifiable {
         self.inLibrary = inLibrary
         self.overview = overview
     }
-}
-
-public extension DiscoverShow {
-  func toSwift() -> SwiftShow {
-    .init(
-      tmdbId: tmdbId,
-      title: title,
-      posterUrl: posterImageUrl,
-      backdropUrl: nil,
-      inLibrary: inLibrary,
-      overview: overView
-    )
-  }
 }
